atualizacoes

+++ petshow-core/src/main/java/br/com/petshow/role/VendaRole.java
@@ -1,13 +1,18 @@
 package br.com.petshow.role;
 
+import java.util.Date;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import br.com.petshow.dao.VendaDAO;
 import br.com.petshow.exceptions.ExceptionNotFoundRecord;
 import br.com.petshow.exceptions.ExceptionValidation;
+import br.com.petshow.model.Anuncio;
 import br.com.petshow.model.Usuario;
 import br.com.petshow.model.Venda;
+import br.com.petshow.util.ValidationUtil;
 /**
  * 
  * @author antoniorafael
@@ -21,25 +26,77 @@ public class VendaRole extends SuperClassRole<Venda> {
 
 	public Venda insert(Venda entidade) throws ExceptionValidation{
 		
+		entidade.setDataCadastro(new Date());
+		entidade.setDataVencimento(entidade.getDataCadastro());
+		
+		if(!ValidationUtil.isCampoComValor(entidade.getDescResumida())){
+			throw new ExceptionValidation("O campo de nome não foi informado!");
+		}
+		if(!ValidationUtil.isCampoComValor(entidade.getDescricao())){
+			throw new ExceptionValidation("O campo de descrição não foi informado!");
+		}
+		if(!ValidationUtil.isCampoComValor(entidade.getDataVencimento())){
+			throw new ExceptionValidation("A data de vencimento não foi informada!");
+		}
+		
+
 		return (Venda) this.vendaDAO.insert(entidade);
 	}
 
 	
 	public void delete(long codigo) throws ExceptionValidation,ExceptionNotFoundRecord {
-	
+		
+		if(!ValidationUtil.isCampoComValor(codigo)){
+			throw new ExceptionValidation("O codigo não foi informado!");
+		}
 		 this.vendaDAO.delete(codigo);
 	}
 
 	
 	public Venda update(Venda entidade) throws ExceptionValidation{
-	
+
+		if(!ValidationUtil.isCampoComValor(entidade.getId())){
+			throw new ExceptionValidation("O codigo não foi informado!");
+		}
+		
+
+		Venda venda = vendaDAO.find(entidade.getId());
+		if(venda==null){
+			throw new ExceptionValidation("Venda informada não existe!");
+		}
+		
+		
+		
+		
+		if(!venda.equals(entidade.getDataCadastro())){
+			throw new ExceptionValidation("Data de cadastro não pode ser alterada!");
+		}
+		if(!ValidationUtil.isCampoComValor(entidade.getDescResumida())){
+			throw new ExceptionValidation("O campo de nome não foi informado!");
+		}
+		if(!ValidationUtil.isCampoComValor(entidade.getDescricao())){
+			throw new ExceptionValidation("O campo de descrição não foi informado!");
+		}
+		if(!ValidationUtil.isCampoComValor(entidade.getDataVencimento())){
+			throw new ExceptionValidation("A data de vencimento não foi informada!");
+		}
 		return (Venda) this.vendaDAO.update(entidade);
 	}
 
 	
 	public Venda find(long codigo) throws ExceptionValidation {
 	
+		if(!ValidationUtil.isCampoComValor(codigo)){
+			throw new ExceptionValidation("O codigo não foi informado!");
+		}
 		return this.vendaDAO.find(codigo);
+	}
+	
+	public List<Venda> consultaPorUsuario(long id)  throws  ExceptionValidation{
+		if(!ValidationUtil.isCampoComValor(id)){
+			throw new ExceptionValidation("O código não foi informado!");
+		}
+		return this.vendaDAO.consultaPorUsuario(id);
 	}
